plugin updates

+++ wp-content/plugins/admin-columns-pro/classes/Search/SegmentRepository/File.php
@@ -27,17 +27,16 @@ final class File implements SegmentRepositoryWritable
 {
 
     use OpCacheInvalidateTrait;
-    use KeyGeneratorTrait;
 
     private const SUFFIX = '_segments';
 
-    private $directory;
+    private Directory $directory;
 
-    private $decoder_factory;
+    private AbstractDecoderFactory $decoder_factory;
 
-    private $encoder_factory;
+    private EncoderFactory $encoder_factory;
 
-    private $serializer;
+    private Serializer $serializer;
 
     public function __construct(
         Directory $directory,
@@ -62,12 +61,7 @@ final class File implements SegmentRepositoryWritable
         return null;
     }
 
-    /**
-     * @param ListScreenId|null $list_screen_id
-     *
-     * @return array
-     */
-    private function load_segments(ListScreenId $list_screen_id = null): array
+    private function load_segments(?ListScreenId $list_screen_id = null): array
     {
         $segments = [];
 
@@ -98,7 +92,7 @@ final class File implements SegmentRepositoryWritable
 
             $encoded_data = require($file->getRealPath());
 
-            if ( ! $this->decoder_factory->can_create($encoded_data)) {
+            if ( ! $this->decoder_factory->supports($encoded_data)) {
                 continue;
             }
 
@@ -120,7 +114,7 @@ final class File implements SegmentRepositoryWritable
         return $segments;
     }
 
-    public function find_all(ListScreenId $list_screen_id = null, Sort $sort = null): SegmentCollection
+    public function find_all(?ListScreenId $list_screen_id = null, ?Sort $sort = null): SegmentCollection
     {
         if (null === $sort) {
             $sort = new Sort\Name();
@@ -137,13 +131,13 @@ final class File implements SegmentRepositoryWritable
 
     public function find_all_personal(
         int $user_id,
-        ListScreenId $list_screen_id = null,
-        Sort $sort = null
+        ?ListScreenId $list_screen_id = null,
+        ?Sort $sort = null
     ): SegmentCollection {
         return new SegmentCollection();
     }
 
-    public function find_all_shared(ListScreenId $list_screen_id = null, Sort $sort = null): SegmentCollection
+    public function find_all_shared(?ListScreenId $list_screen_id = null, ?Sort $sort = null): SegmentCollection
     {
         return $this->find_all($list_screen_id, $sort);
     }
@@ -167,6 +161,9 @@ final class File implements SegmentRepositoryWritable
         if ($this->find($segment->get_key())) {
             throw FailedToSaveSegmentException::from_duplicate_key($segment->get_key());
         }
+
+        // Mark as modified
+        $segment = $segment->with_modified_now();
 
         $file = $this->get_file_name($segment->get_list_id());
 
@@ -205,7 +202,7 @@ final class File implements SegmentRepositoryWritable
         $segments->remove($key);
 
         if ($segments->count() < 1) {
-            $this->delete_file($file, $list_id);
+            $this->delete_file($file);
 
             return;
         }
@@ -227,7 +224,7 @@ final class File implements SegmentRepositoryWritable
         );
 
         if (file_exists($file)) {
-            $this->delete_file($file, $list_screen_id);
+            $this->delete_file($file);
         }
     }
 
@@ -247,7 +244,7 @@ final class File implements SegmentRepositoryWritable
         if ($info->isFile()) {
             $encoded_data = require($info->getRealPath());
 
-            if (is_array($encoded_data) && $this->decoder_factory->can_create($encoded_data)) {
+            if (is_array($encoded_data) && $this->decoder_factory->supports($encoded_data)) {
                 $decoder = $this->decoder_factory->create($encoded_data);
 
                 if ($decoder instanceof Decoder\SegmentsDecoder && $decoder->has_segments()) {
@@ -275,7 +272,7 @@ final class File implements SegmentRepositoryWritable
         );
 
         if ($result === false) {
-            throw FileNotWritableException::from_saving_segment($list_screen_id);
+            throw FileNotWritableException::for_file($file);
         }
 
         $this->opcache_invalidate($file);
@@ -284,14 +281,18 @@ final class File implements SegmentRepositoryWritable
     /**
      * @throws FileNotWritableException
      */
-    private function delete_file(string $file, ListScreenId $list_screen_id): void
+    private function delete_file(string $file): void
     {
         $this->opcache_invalidate($file);
+
+        if ( ! file_exists($file)) {
+            return;
+        }
 
         $result = unlink($file);
 
         if ($result === false) {
-            throw FileNotWritableException::from_removing_segment($list_screen_id);
+            throw FileNotWritableException::for_file($file);
         }
     }
